added a workout interval timer ; cleanup
The HIIT workout timer lets you do interval exercises. The exercise
comprises blocks of workout intervals separated by rest intervals.
Within each block, the user exercises for a number of ON/OFF periods.
By default, a block is 12 on/off periods, on=30s, off=10s, rest=2m
between blocks, and the workout is 3 blocks, for a total of 32 minutes
(12 * 40s * 3 + 4 * 120s).

I added hooks to make this selectable as a module much like the egg
timer.

TODO: allow the user to set up length and number of intervals and
blocks other than by changing the code.

In addition, I cleaned up some code. In the data display code,
there was an issue (at least, when day of week is turned off)
where alternating between date and year is broken. Also, the
code fails to compile when the press-backlight option is enabled
because there is some code that's not #ifndef'd out in ports.c.

Finally, removed ^M characters from eggtimer and phase_clock code.